double

+++ T2_POO/ContratoDeJogos/src/dados/Contrato.java
@@ -58,7 +58,7 @@ da categoria do jogo e da forma de pagamento:*/
 
     public void calculaValorFinal(){
         double  result = 0;
-        int valorDiario = jogo.getValorDiario();
+        double valorDiario = jogo.getValorDiario();
         Categoria categoriaCon = jogo.getCategoria();
 
         if(categoriaCon.getExtenso().equals("AVENTURA")){
